+++ Demo/VComponentsDemo/Scenes/Demos/Lists/VStaticListDemoView.swift
@@ -28,7 +28,7 @@ struct VStaticListDemoView: View {
 
     // MARK: Body
     var body: some View {
-        DemoView(component: component, settingsSections: settings)
+        DemoView(paddedEdges: .vertical, component: component, settingsSections: settings)
             .standardNavigationTitle(Self.navBarTitle)
     }